pyproject.toml
[project]
name = "image-classifier"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "aesthetic-predictor-v2-5",
    "torch>=2.0",
    "transformers>=4.30,<5",
    "Pillow",
    "rich",
]

[project.scripts]
classify = "image_classifier.main:main"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.15.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/image_classifier"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=image_classifier --cov-report=term-missing"

[tool.coverage.run]
source = ["src/image_classifier"]
# main.py is the CLI entry point requiring GPU + real files; not testable in CI
omit = ["src/image_classifier/main.py"]
